extracted APIs
